% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Basic-class}
\alias{Basic-class}
\title{Class "Basic"}
\description{
A class to represent basic metrics
}
\section{Slots}{

\describe{
\item{\code{summary_data}}{summary data}

\item{\code{reads_count}}{read counts}

\item{\code{clonotype_count}}{clonotype count}

\item{\code{mean_freq}}{mean frequency}

\item{\code{geomean_freq}}{geometric mean of clonotype frequency}

\item{\code{mean_cdr3nt_len}}{mean CDR3 nucleotide length}

\item{\code{convergence}}{mean of the unique CDR3 count in each sample}

\item{\code{spectratype}}{spectratype data}
}}

