% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Clonality-class}
\alias{Clonality-class}
\title{Class "Clonality"}
\description{
A class to represent clonality metrics
}
\section{Slots}{

\describe{
\item{\code{most_clonotype}}{most frequent clonotype}

\item{\code{least_clonotype}}{least frequent clonotype}

\item{\code{pielou}}{1-Pielou index}

\item{\code{clonal_prop}}{clonal proportion}

\item{\code{abundance}}{relative abundance (in all clonotypes)}

\item{\code{abundance_top}}{relative abundance (in top 100 clonotypes)}

\item{\code{abundance_rare}}{relative abundance (in rare clonotypes)}
}}

