% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Diversity-class}
\alias{Diversity-class}
\title{Class "Diversity"}
\description{
A class to represent diversity data.
}
\section{Slots}{

\describe{
\item{\code{shannon_index}}{A numeric value representing the Shannon index.}

\item{\code{gini_index}}{A numeric value representing the Gini index.}
}}

