% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get pyTCR metrics of a study}
\usage{
get_metrics(id, type)
}
\arguments{
\item{id}{A character string specifying the ID of the data for which metrics are to be retrieved.}

\item{type}{An enumeration specifying the type of metrics to retrieve. Possible values are:
\itemize{
\item "basic" - Basic metrics.
\item "metadata" - Metadata.
\item "clonality" - Clonality metrics.
\item "segment" - Segment usage metrics.
\item "diversity" - Diversity metrics.
\item "motif" - Motif metrics.
}}
}
\value{
A data frame containing the requested metrics.
}
\description{
Retrieve metrics based on provided ID and type.
}
\examples{
\dontrun{
# Retrieve basic metrics for ID "sample1"
basic_metrics <- reTCR::get_metrics(id="PRJNA473147", type="basic")
}

}
