% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study.R
\name{get_study}
\alias{get_study}
\title{Get pyTCR data of a project}
\usage{
get_study(id, attribute_col)
}
\arguments{
\item{id}{Character. Project ID.}

\item{attribute_col}{Character. Attribute column name.}
}
\value{
An object of class \code{reTCRProj}.
}
\description{
Retrieve the pyTCR data of a project by its ID.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id="PRJNA473147", attribute_col="cmv_status")
}
}
