% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-segment.R
\name{get_top_n_clonotypes}
\alias{get_top_n_clonotypes}
\title{Plot top n clonotypes}
\usage{
get_top_n_clonotypes(data, n, attr_col)
}
\arguments{
\item{data}{A data frame of the main data.}

\item{n}{Integer. Number of top clonotypes to select.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A data frame of top n clonotypes.
}
\description{
This function plots the top n clonotypes.
}
\examples{
\dontrun{
proj <- reTCR::get_study("PRJNA473147", "cmv_status")
reTCR::get_top_n_clonotypes(proj@data, 10, "cmv_status")
}
}
