% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Motif-class}
\alias{Motif-class}
\title{Class "Motif"}
\description{
A class to represent motif metrics
}
\section{Slots}{

\describe{
\item{\code{aa_spectra}}{Amino acid spectratype}

\item{\code{aa_max_spectra}}{Amino acid max spectratype}

\item{\code{aa_motif_count}}{Amino acid motif count table}

\item{\code{aa_most_motif}}{Most abundant amino acid motif per sample}
}}

