% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-clonality-plot.R
\name{plot_clonal_prop_per_group}
\alias{plot_clonal_prop_per_group}
\title{Plot clonal proportion per group}
\usage{
plot_clonal_prop_per_group(data, attr_col)
}
\arguments{
\item{data}{A data frame of clonal proportion per group.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A ggplot2 plot of clonal proportion per group.
}
\description{
This function plots the clonal proportion per group.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id = "PRJNA473147", attr_col = "cmv_status")
clonal_prop <- proj@clonality@clonal_prop
reTCR::plot_clonal_prop_per_group(clonal_prop, "cmv_status")
}
}
