% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clonal_prop_per_sample.R
\name{plot_clonal_prop_per_sample}
\alias{plot_clonal_prop_per_sample}
\title{Plot Clonal Proportion per Sample}
\usage{
plot_clonal_prop_per_sample(data, attr_col)
}
\arguments{
\item{data}{A data frame containing the clonal proportion per sample.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A ggplot2 object representing the clonal proportion per sample plot.
}
\description{
This function plots the clonal proportion per sample.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id="PRJNA473147", attr_col="cmv_status")
clonal_prop <- proj@clonality@clonal_prop
reTCR::plot_clonal_prop_per_sample(clonal_prop, "cmv_status")
}
}
