% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_clonotype_abundance.R
\name{plot_clonotype_abundance}
\alias{plot_clonotype_abundance}
\title{Plot relative abundance for all clonotypes}
\usage{
plot_clonotype_abundance(data)
}
\arguments{
\item{data}{A data frame of relative abundance for all clonotypes.}
}
\value{
A ggplot2 plot of relative abundance for all clonotypes.
}
\description{
This function plots the relative abundance for all clonotypes.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id="PRJNA473147", attr_col="cmv_status")
reTCR::plot_clonotype_abundance(proj@clonality@abundance)
}
}
