% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity_index.R
\name{plot_diversity_index}
\alias{plot_diversity_index}
\title{Plot diversity index metrics}
\usage{
plot_diversity_index(data, index, attr_col)
}
\arguments{
\item{data}{A data frame of diversity index metrics.}

\item{index}{Character. Diversity index metric.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A ggplot2 plot of diversity index metrics.
}
\description{
This function plots the diversity index metrics.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id = "PRJNA473147", attr_col = "cmv_status")
reTCR::plot_diversity_index(
  proj@diversity@shannon,
  "shannon_wiener_index",
  "cmv_status"
)
}
}
