% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-motif-plot.R
\name{plot_motifs}
\alias{plot_motifs}
\title{Plot Motif Counts}
\usage{
plot_motifs(data, attr_col, k = 60)
}
\arguments{
\item{data}{A data frame containing the motif counts.}

\item{attr_col}{Character. Attribute column name.}

\item{k}{the minimum count threshold (default: 60).}
}
\value{
A ggplot2 object representing the motif counts plot.
}
\description{
This function plots the motif counts.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id="PRJNA473147", attr_col="cmv_status")
reTCR::plot_motifs(proj@motif@aa_motif_count, "cmv_status", 50)
}
}
