% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reads_group_abundance.R
\name{plot_reads_group_abundance}
\alias{plot_reads_group_abundance}
\title{Plot reads group abundance for top 100 clonotypes}
\usage{
plot_reads_group_abundance(data)
}
\arguments{
\item{data}{A data frame of relative abundance for top 100 clonotypes.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A ggplot2 plot of reads group abundance for top 100 clonotypes.
}
\description{
This function plots reads group abundance for top 100 clonotypes.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id = "PRJNA473147", attr_col = "cmv_status")
reTCR::plot_reads_group_abundance(proj@clonality@abundance_top)
}
}
