% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_abundance.R
\name{plot_relative_abundance}
\alias{plot_relative_abundance}
\title{Plot relative abundance for all clonotypes}
\usage{
plot_relative_abundance(data)
}
\arguments{
\item{data}{A data frame of relative abundance for all clonotypes.}

\item{attr_col}{Character. Attribute column name.}
}
\value{
A ggplot2 plot of relative abundance for all clonotypes.
}
\description{
This function plots the relative abundance for all clonotypes.
}
\examples{
\dontrun{
proj <- reTCR::get_study(id = "PRJNA473147", attr_col = "cmv_status")
reTCR::plot_relative_abundance(proj@clonality@abundance)
}
}
