% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Diversity-class}
\alias{Diversity-class}
\title{Class "Diversity"}
\description{
A class to represent diversity data.
}
\section{Slots}{

\describe{
\item{\code{summary_data}}{summary data}

\item{\code{shannon}}{Shannon-Wiener indexes}

\item{\code{simpson}}{Simpson indexes}

\item{\code{d50}}{D50 index}

\item{\code{chao1}}{Chao1 estimate}

\item{\code{gini_coeff}}{Gini coefficient}
}}

