% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study.R
\name{get_study}
\alias{get_study}
\title{Get pyTCR data of a project}
\usage{
get_study(id)
}
\arguments{
\item{id}{Project ID.}
}
\value{
Data frame with the project data.
}
\description{
Get pyTCR data of a project
}
\examples{
\dontrun{ 
# Get data with ID "PRJNA473147"
project <- reTCR::get_study(id="PRJNA473147")
}
}
