% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study.R
\name{get_study}
\alias{get_study}
\title{Get pyTCR metrics of a study
Retrieve metrics based on project ID.}
\usage{
get_study(id)
}
\arguments{
\item{id}{A character string specifying the ID of the data for which metrics are to be retrieved.}
}
\value{
A data frame containing the requested metrics.
}
\description{
Get pyTCR metrics of a study
Retrieve metrics based on project ID.
}
\examples{
\dontrun{
# Retrieve project with ID "PRJNA473147"
project <- reTCR::get_study(id="PRJNA473147")
}

}
