% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Motif-class}
\alias{Motif-class}
\title{Class "Motif"}
\description{
A class to represent motif data.
}
\section{Slots}{

\describe{
\item{\code{aa_count}}{A numeric value representing the amino acid count.}

\item{\code{nucl_count}}{A numeric value representing the nucleotide count.}

\item{\code{spectratype}}{A data frame containing spectratype data.}
}}

