% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{Motif-class}
\alias{Motif-class}
\title{Class "Motif"}
\description{
A class to represent motif data
}
\section{Slots}{

\describe{
\item{\code{aa_spectratype}}{aa spectratype data}

\item{\code{aa_max_spectratype}}{aa max spectratype data}

\item{\code{aa_motif_count}}{aa motif count data}
}}

