% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retcr_proj_class.R
\docType{class}
\name{reTCRProj-class}
\alias{reTCRProj-class}
\title{Class "reTCRProj"}
\description{
A class to represent a reTCR project
}
\section{Slots}{

\describe{
\item{\code{data}}{contains the main data for the project.}

\item{\code{basic}}{contains basic analysis information.}
}}

